utils/requirements.txt
requests==2.31.0
python-binance==1.0.19
python-dotenv==1.0.0 